+++ rust/bridge/shared/types/src/node/error.rs
@@ -557,8 +557,9 @@ impl SignalNodeError for libsignal_net::cdsi::LookupError {
 mod registration {
     use libsignal_net::infra::errors::RetryLater;
     use libsignal_net::registration::{
-        CreateSessionError, RequestError, RequestVerificationCodeError, ResumeSessionError,
-        SubmitVerificationError, UpdateSessionError, VerificationCodeNotDeliverable,
+        CreateSessionError, RegisterAccountError, RegistrationLock, RequestError,
+        RequestVerificationCodeError, ResumeSessionError, SubmitVerificationError,
+        UpdateSessionError, VerificationCodeNotDeliverable,
     };
 
     use super::*;
@@ -604,6 +605,9 @@ mod registration {
         NotReadyForVerification,
         VerificationSendFailed,
         VerificationNotDeliverable(VerificationCodeNotDeliverable),
+        RegistrationLock(RegistrationLock),
+        RecoveryVerificationFailed,
+        DeviceTransferPossibleNotSkipped,
     }
 
     impl SignalNodeError for BridgedErrorVariant {
@@ -631,6 +635,15 @@ mod registration {
                 }
                 BridgedErrorVariant::VerificationNotDeliverable(_not_deliverable) => {
                     "the verification code could not be delivered"
+                }
+                BridgedErrorVariant::RecoveryVerificationFailed => {
+                    "the recovery password was not accepted"
+                }
+                BridgedErrorVariant::DeviceTransferPossibleNotSkipped => {
+                    "device transfer is possible but wasn't explicitly skipped"
+                }
+                BridgedErrorVariant::RegistrationLock(_registration_lock) => {
+                    "registration is locked"
                 }
             };
             new_js_error(
@@ -697,6 +710,23 @@ mod registration {
                 SubmitVerificationError::SessionNotFound => Self::SessionNotFound,
                 SubmitVerificationError::NotReadyForVerification => Self::NotReadyForVerification,
                 SubmitVerificationError::RetryLater(retry_later) => Self::RetryLater(retry_later),
+            }
+        }
+    }
+
+    impl From<RegisterAccountError> for BridgedErrorVariant {
+        fn from(value: RegisterAccountError) -> Self {
+            match value {
+                RegisterAccountError::DeviceTransferIsPossibleButNotSkipped => {
+                    Self::DeviceTransferPossibleNotSkipped
+                }
+                RegisterAccountError::RegistrationRecoveryVerificationFailed => {
+                    Self::RecoveryVerificationFailed
+                }
+                RegisterAccountError::RegistrationLock(registration_lock) => {
+                    Self::RegistrationLock(registration_lock)
+                }
+                RegisterAccountError::RetryLater(retry_later) => Self::RetryLater(retry_later),
             }
         }
     }